вращение/зум пальцем

+++ app/src/main/java/com/example/hzkto/ball/sphere/Point3D.java
@@ -5,9 +5,9 @@ package com.example.hzkto.ball.sphere;
  */
 
 public class Point3D {
-    public float x, y, z;
+    public double x, y, z;
 
-    public Point3D(float x, float y, float z) {
+    public Point3D(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
